move clientconfig to package root
